autotests_gologin/requirements.txt
pytest==7.2.1
pytest-html==3.2.0
pytest-random-order==1.1.0
pytest-dependency==0.5.1
pytest-ordering==0.6
pytest-metadata==2.0.4
pytest-rerunfailures==11.1.1
pytest-testconfig==0.2.0
allure-pytest==2.12.0
requests==2.28.2




